backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "docassemble-editor-backend"
version = "0.1.0"
description = "API layer for the Docassemble visual interview builder"
authors = [
  { name = "Code Lab" }
]
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.9.0",
  "ruamel.yaml>=0.18.10",
  "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.3.0", "httpx[http2]>=0.27.0"]

[project.scripts]
docassemble-editor-api = "app.main:run"
